fix(diarization): auto-resolve directory model sources in native bridge and add unit test
