+++ src/diarization/__tests__/createDiarization.test.ts
@@ -2,6 +2,7 @@ jest.mock('../../NativeSherpaOnnx', () => ({
   __esModule: true,
   default: {
     detectDiarizationModel: jest.fn(),
+    detectSpeakerEmbeddingModel: jest.fn(),
     initializeDiarization: jest.fn(),
     unloadDiarization: jest.fn(),
     diarizeOffline: jest.fn(),
@@ -64,6 +65,8 @@ import { createDiarization, DiarizationErrorCode } from '../index';
 
 describe('createDiarization', () => {
   const native = SherpaOnnx as unknown as {
+    detectDiarizationModel: jest.Mock;
+    detectSpeakerEmbeddingModel: jest.Mock;
     initializeDiarization: jest.Mock;
     unloadDiarization: jest.Mock;
     diarizeOffline: jest.Mock;
@@ -181,5 +184,33 @@ describe('createDiarization', () => {
     });
     // abort may fire after resolve depending on timing; ensure cancel is registered
     expect(native.cancelDiarization).toBeDefined();
+  });
+
+  it('resolves model directory FileSource to detected onnx model files', async () => {
+    native.detectDiarizationModel.mockResolvedValue({
+      success: true,
+      paths: { model: '/models/seg-dir/model.onnx' },
+    });
+    native.detectSpeakerEmbeddingModel.mockResolvedValue({
+      success: true,
+      paths: { model: '/models/emb-dir/3dspeaker.onnx' },
+    });
+
+    await createDiarization({
+      segmentation: {
+        modelSource: { kind: 'fs', path: '/models/seg-dir' },
+      },
+      embedding: {
+        modelSource: { kind: 'fs', path: '/models/emb-dir' },
+      },
+    });
+
+    expect(native.initializeDiarization).toHaveBeenCalledWith(
+      expect.any(String),
+      expect.objectContaining({
+        segmentationModel: '/models/seg-dir/model.onnx',
+        embeddingModel: '/models/emb-dir/3dspeaker.onnx',
+      })
+    );
   });
 });
